% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{add_a_leaf_to_foliox}
\alias{add_a_leaf_to_foliox}
\title{Add a Leaf to a \code{foliox}}
\usage{
add_a_leaf_to_foliox(headv, foliox, leafname, bhat, chat)
}
\arguments{
\item{headv}{head vertex of edge to attach to}

\item{foliox}{phylogenetic leaf address object (foliox)}

\item{leafname}{label of taxon at the leaf to be attached}

\item{bhat}{length of pendant edge to the leaf}

\item{chat}{length of subdivided edge from attachment point to tail of edge}
}
\value{
foliox object
}
\description{
Attach a leaf to a specified edge in a tree encoded as \code{foliox} object. The
function requires details about the subdivided edge. Note that this
adds a single leaf to the tree.
}
\examples{
  add_a_leaf_to_foliox(10, egfoliox_initial, "new_taxon", 0.02, 0.03)

}
\seealso{
\code{\link{place_seqs}}
}
\keyword{foliox}
\keyword{leaf}
