% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{distance_between}
\alias{distance_between}
\title{Compute Distance Between Vertices in \code{foliox}}
\usage{
distance_between(v1, v2, foliox)
}
\arguments{
\item{v1}{the first vertex}

\item{v2}{the second vertex}

\item{foliox}{phylogenetic leaf address object (foliox)}
}
\value{
distance computed from edge weights
}
\description{
Compute the distance between two vertices in a \code{foliox} object. A vertex
can be a leaf (tip).
}
\examples{
 distance_between(12,8,egfoliox_initial)

}
\seealso{
\code{\link{distance_to_vertex}, \link{divergence_point}}
}
\keyword{distance}
\keyword{foliox}
