% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{distance_to_vertex}
\alias{distance_to_vertex}
\title{Distance From Reference Node to a Vertex}
\usage{
distance_to_vertex(vert, foliox)
}
\arguments{
\item{vert}{vertex label in foliox}

\item{foliox}{phylogenetic leaf address object (foliox)}
}
\value{
distance computed from edge weights
}
\description{
Find distance from the reference node to any vertex, including a leaf.
}
\examples{
 distance_to_vertex(9,egfoliox_initial)
}
\seealso{
\code{\link{distance_between}, \link{divergence_point}}
}
\keyword{distance}
\keyword{foliox}
\keyword{vertex}
