% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{divergence_point}
\alias{divergence_point}
\title{Divergence point between two vertices}
\usage{
divergence_point(v1, v2, foliox)
}
\arguments{
\item{v1}{the first vertex}

\item{v2}{the second vertex}

\item{foliox}{phylogenetic leaf address object (foliox)}
}
\value{
vertex label from foliox
}
\description{
Find divergence point between two vertices in a \code{foliox} object. This is
the most recent common ancestor (MRCA) if the reference node is the
root of the tree.
}
\examples{
 divergence_point(14,20,egfoliox_initial)

}
\seealso{
\code{\link{distance_between}, \link{distance_to_vertex}}
}
\keyword{divergence}
\keyword{foliox}
\keyword{vertex}
