% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\docType{data}
\name{egfoliox}
\alias{egfoliox}
\alias{egfoliox_subdivs}
\alias{egfoliox_polytomy}
\alias{egfoliox_initial}
\alias{egfoliox_updated}
\title{Example \code{foliox} Data}
\format{
An object of class \code{foliox} of length 4.

An object of class \code{foliox} of length 4.

An object of class \code{foliox} of length 4.

An object of class \code{foliox} of length 4.
}
\usage{
egfoliox_subdivs

egfoliox_polytomy

egfoliox_initial

egfoliox_updated
}
\description{
Hypothetical data in foliox format for illustrating the use of functions
in the \code{splade} package. \code{egfoliox_subdivs} (\code{foliox} with subdivided edges);
\code{egfoliox_polytomy} (\code{foliox} with polytomies); \code{egfoliox_initial} (an initial
\code{foliox} before more leaves are added); \code{egfoliox_updated} (an updated version
of \code{egfoliox_initial} with leaves added).
}
\keyword{datasets}
