% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{foliox2apex}
\alias{foliox2apex}
\title{Convert \code{foliox} to ape's \code{phylo} Format with Extra Information}
\usage{
foliox2apex(foliox, address.for.tips = FALSE)
}
\arguments{
\item{foliox}{phylogenetic leaf address object (foliox)}

\item{address.for.tips}{whether to use addresses as tip labels}
}
\value{
A list including
\describe{
\item{\code{ape}       }{ tree as phylo class}
\item{\code{head.label}}{ labels for head vertices}
\item{\code{intr.label}}{ interior node labels}
\item{\code{tip.nums}  }{ tip numbers from the foliox}
}
}
\description{
Convert a \code{foliox} object to a list containing a tree in ape's \code{phylo} class
and additional information: labels for head vertices, interior node
labels, and tip numbers from the \code{foliox}.
}
\examples{
 apex <- foliox2apex(egfoliox_initial)
 apex$ape   # this is the tree as phylo object
 apex       # tree and extra other information

}
\seealso{
\code{\link{foliox2phylo}, \link{phylo2foliox},
   \link{newick2foliox}, \link{egfoliox}}
}
\keyword{foliox}
\keyword{phylo}
