% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{foliox2phylo}
\alias{foliox2phylo}
\title{Convert \code{foliox} Object to ape's \code{phylo} Format}
\usage{
foliox2phylo(foliox, address.for.tips = FALSE)
}
\arguments{
\item{foliox}{phylogenetic leaf address object (foliox)}

\item{address.for.tips}{whether to use addresses as tip labels}
}
\value{
a tree as phylo
}
\description{
Convert phylogenetic leaf address (\code{foliox}) object to a tree in ape's
\code{phylo} format.
}
\examples{
  foliox2phylo(egfoliox_initial)
}
\seealso{
\code{\link{foliox2apex}, \link{phylo2foliox}, \link{newick2foliox}, \link{egfoliox}}
}
\keyword{foliox}
\keyword{phylo}
