% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{foliox_clade}
\alias{foliox_clade}
\title{Create a New \code{foliox} From a Clade Within a \code{foliox}}
\usage{
foliox_clade(foliox, headv, remove.prefix = TRUE)
}
\arguments{
\item{foliox}{phylogenetic leaf address object (foliox)}

\item{headv}{label of head vertex at the clade to be extracted}

\item{remove.prefix}{exclude the prefix from addresses}
}
\value{
foliox object
}
\description{
Extract a clade using the head vertex above it; return clade as a new
\code{foliox}.
}
\examples{
 plot(egfoliox_initial) 
 fc <- foliox_clade(egfoliox_initial, 10) # extract clade with head vertex 10
 plot(fc)
}
\seealso{
\code{\link{foliox_subset}, \link{egfoliox}}
}
\keyword{clade}
\keyword{foliox}
\keyword{subtree}
