% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{foliox_subset}
\alias{foliox_subset}
\title{Create a \code{foliox} From a Subset of Addresses}
\usage{
foliox_subset(foliox, subset)
}
\arguments{
\item{foliox}{phylogenetic leaf address object (foliox)}

\item{subset}{vector of indices of addresses to extract or vector of names of taxa}
}
\value{
foliox object
}
\description{
This takes a \code{foliox} and the indices of a subset of addresses to extract
from the \code{foliox}, and creates a new \code{foliox} from it. For compatibility with
the original \code{foliox} the subset carries interior nodes that are unused
in the subset.
}
\examples{
 egfoliox_initial  # inspect this example foliox
 foliox_subset(egfoliox_initial, c(1,3,4)) # rebuild foliox with three addresses
 foliox_subset(egfoliox_initial, c("t2","t18","t22","t23"))
}
\seealso{
\code{\link{foliox_clade}, \link{egfoliox}}
}
\keyword{foliox}
\keyword{subtree}
