% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{newick2foliox}
\alias{newick2foliox}
\title{Read a \code{newick} File and Convert to \code{foliox}}
\usage{
newick2foliox(tree.file = "", text = "")
}
\arguments{
\item{tree.file}{file containing newick format tree}

\item{text}{newick as text}
}
\value{
foliox object
}
\description{
Read a tree from a file in newick format and convert the tree to foliox.
}
\examples{
 newick.tree.file <- system.file("extdata","tree-inferred-initial.newick",
                              package="splade")
 foliox <- newick2foliox(newick.tree.file)
 summary(foliox)
 filenamestem <- tempfile("tree-as-foliox")
 # filenamestem <- "tree-as-foliox"  # to save local copy 
 write_foliox(foliox, filenamestem) # write to files

 newick_tree <- "( (1,2), ((3, 4), 5) );"
 newick2foliox(text=newick_tree)

}
\seealso{
\code{\link{foliox2apex}, \link{phylo2foliox}, \link{foliox2phylo}}
}
\keyword{foliox}
\keyword{newick}
