% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{phylo2foliox}
\alias{phylo2foliox}
\title{Convert from \code{phylo} to \code{foliox}}
\usage{
phylo2foliox(ape.tree, elabels = NULL)
}
\arguments{
\item{ape.tree}{tree as phylo}

\item{elabels}{vector for labelling edges; if none supplied, use
numbering in ape's phylo}
}
\value{
foliox object
}
\description{
Convert a tree in \code{phylo} format (ape package) to a phylogenetic leaf
address (\code{foliox}) object.
}
\examples{

 ape.tree <- ape::rtree(16)  # make a random tree using ape
 foliox <- phylo2foliox(ape.tree)
 plot(foliox)

 # Go from newick to phylo to foliox, then back to phylo
 newick_tree <- "( 1, (2, 8, 9, 10, (((3, 4, 7), 5), 6) ) );"
 atree <- ape::read.tree(text=newick_tree) # convert to phylo format
 foliox <- phylo2foliox(atree)  # convert to foliox
 foliox
 tree <- foliox2phylo(foliox)  # convert back to phylo
 plot(tree)

}
\seealso{
\code{\link{foliox2phylo}, \link{foliox2apex}, \link{newick2foliox}, \link{egfoliox}}
}
\keyword{foliox}
\keyword{phylo}
