% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{place_seqs}
\alias{place_seqs}
\title{Add New Sequences to Enlarge a Tree}
\usage{
place_seqs(
  seq.file,
  tree.file,
  new.seq.file,
  seq.format = "sequential",
  tree.format = "foliox"
)
}
\arguments{
\item{seq.file}{file containing sequences in phylip format}

\item{tree.file}{file containing newick or foliox tree inferred
from the seq.file; if foliox, use filename stem}

\item{new.seq.file}{file containing new sequences to be added to the
current data in seq.file}

\item{seq.format}{format of the sequence files, e.g. "sequential", or "fasta"}

\item{tree.format}{format of the input tree, i.e., "foliox" or "newick"}
}
\value{
foliox object
}
\description{
Enlarge a tree by adding new leaves using new sequences to be appended
to an existing set of sequences. The leaf placement algorithm uses the
Fitch-Margoliash criterion. The input tree can be either newick or
foliox. Return a \code{foliox}.
}
\examples{

 # example (simulated) sequences in sequential format 
 sequences.initial <- system.file("extdata","sequences-initial.txt",
                                   package="splade")
 tree.initial <- system.file("extdata","tree-inferred-initial.newick",
                              package="splade")
 sequences.additional <- system.file("extdata","sequences-additional.txt",
                                     package="splade")

 newfoliox <- place_seqs(seq.file=sequences.initial,
                         tree.file=tree.initial,
                         new.seq.file=sequences.additional,
                         tree.format="newick")

 folioxname <- tempfile("updated-foliox")
 # folioxname <- "updated-foliox"  # to save local copy
 write_foliox(newfoliox, folioxname)  # save it 
 plot(newfoliox)

}
\seealso{
\code{\link{add_a_leaf_to_foliox}}
}
\keyword{newick}
\keyword{sequences}
