% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{plot.foliox}
\alias{plot.foliox}
\title{A Plot Method for Class \code{foliox}}
\usage{
\method{plot}{foliox}(
  x,
  ...,
  address.for.tips = TRUE,
  foliox.node.lab = TRUE,
  foliox.tip.lab = FALSE,
  node.cex = 0.8,
  tip.cex = 0.8
)
}
\arguments{
\item{x}{phylogenetic leaf address object (foliox)}

\item{...}{additional parameters to pass to ape's plot function}

\item{address.for.tips}{whether to use addresses as tip labels}

\item{foliox.node.lab}{whether to add interior node labels from foliox}

\item{foliox.tip.lab}{whether to add tip labels from foliox}

\item{node.cex}{cex factor for foliox nodes}

\item{tip.cex}{cex factor for foliox tips}
}
\description{
Plot a \code{foliox} by converting to \code{phylo} then using ape to render the tree.
}
\examples{

plot(egfoliox_polytomy, type="cladogram", label.offset=0.005,
    show.tip.label=FALSE, foliox.tip.lab=TRUE, foliox.node.lab=FALSE)

plot(egfoliox_subdivs, type="unrooted")
plot(egfoliox_initial, node.cex=0.75)
plot(egfoliox_updated, node.cex=0.75)

}
\seealso{
The plot function in ape: {ape::plot.phylo}
}
\keyword{foliox}
