% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{print.foliox}
\alias{print.foliox}
\title{Print Method for Class \code{foliox}}
\usage{
\method{print}{foliox}(x, ...)
}
\arguments{
\item{x}{phylogenetic leaf address object (foliox)}

\item{...}{other print parameters}
}
\description{
Print \code{foliox} object.
}
\examples{
 print(egfoliox_polytomy) 
 print(egfoliox_subdivs)
 print(egfoliox_initial)
 print(egfoliox_updated)

}
\keyword{foliox}
