% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{read_foliox}
\alias{read_foliox}
\title{Read \code{foliox} from two files}
\usage{
read_foliox(filenamestem)
}
\arguments{
\item{filenamestem}{stem or basename of file (reads in two files)}
}
\value{
foliox object
}
\description{
Read a \code{foliox} object from two files. The first stores names and
addresses, the second stores head vertices and weights, i.e.,
distances.
}
\examples{
 filenamestem <- tempfile("initfoliox")
 # filenamestem <- "initfoliox"  # to save local copy 
 write_foliox(egfoliox_initial, filenamestem)
 read_foliox(filenamestem)

}
\seealso{
\code{\link{write_foliox}}
}
\keyword{foliox}
\keyword{read}
