% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{summary.foliox}
\alias{summary.foliox}
\title{Summary Method for Class \code{foliox}}
\usage{
\method{summary}{foliox}(object, ...)
}
\arguments{
\item{object}{phylogenetic leaf address object (foliox)}

\item{...}{other parameters for summary of weights}
}
\description{
Summarise \code{foliox} object.
}
\examples{
 summary(egfoliox_polytomy)
 summary(egfoliox_subdivs)
 summary(egfoliox_initial)
 summary(egfoliox_updated)

}
\keyword{foliox}
