% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splade.R
\name{write_foliox}
\alias{write_foliox}
\title{Write \code{foliox} to Two Files}
\usage{
write_foliox(foliox, filenamestem = "")
}
\arguments{
\item{foliox}{phylogenetic leaf address object (foliox)}

\item{filenamestem}{stem of filename for the two files}
}
\description{
Save a \code{foliox} object to two files with names starting with string
supplied through \code{filenamestem}. The first file stores names and
addresses, the second stores head vertices and weights (i.e.,
distances). The two file extensions are \code{.foladdress} and
\code{.folweight} respectively.
}
\examples{
 filenamestem <- tempfile("initfoliox")
 # filenamestem <- "initfoliox"  # to save local copy 
 write_foliox(egfoliox_initial, filenamestem)

}
\seealso{
\code{\link{read_foliox}}
}
\keyword{foliox}
\keyword{write}
